AbstractNetworkFilter moved to abstract_network.rs.
