+++ src/filters/mod.rs
@@ -1,4 +1,5 @@
 //! Contains representations and standalone behaviors of individual filter rules.
 
+mod abstract_network;
 pub mod cosmetic;
 pub mod network;
